Refactor map exports and enhance ZoomLevelControl: update export paths, add Theme support for DSFR theme.
