+++ src/components/ZoomLevelControl/ZoomLevelControl.css
@@ -2,4 +2,9 @@
   padding: 4px 8px;
   font-weight: 700;
   color: #000;
-} 
+}
+
+/* Thème DSFR */
+[data-theme="dsfr"].maplibregl-ctrl-zoom-level {
+  font-family: 'Marianne', Arial, sans-serif;
+}
