+++ src/index.ts
@@ -1,3 +1,4 @@
-export * from './map/types';
-export * from './map/maps';
-export { ZoomLevelControl } from './components/ZoomLevelControl/ZoomLevelControl'; 
+export * from './maps/types';
+export * from './maps/maps';
+export { ZoomLevelControl } from './components/ZoomLevelControl/ZoomLevelControl';
+export { Theme } from './themes/manager';
