+++ src/maps/overlays/level-curves/color.layers.json
