+++ src/maps/overlays/level-curves/common.json
